% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_net.R
\name{add_correlation_to_coexpression_modules}
\alias{add_correlation_to_coexpression_modules}
\title{Title}
\usage{
add_correlation_to_coexpression_modules(GRANetObject, mask_dropouts = FALSE)
}
\arguments{
\item{threads}{}
}
\value{

}
\description{
Title
}
