% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssRegulons.R
\name{add_motifs_position_from_ArchR}
\alias{add_motifs_position_from_ArchR}
\title{Title}
\usage{
add_motifs_position_from_ArchR(
  GRANetObject,
  ArchRProjectObj,
  cssClusterArchR,
  promoter_size = 5000,
  cellsWithPeak = 0.01,
  threads = 1
)
}
\arguments{
\item{cellsWithPeak}{Minimum percentage of cells in a cluster, that are
required to have at least one insertion in a peak to keep it}

\item{threads}{}
}
\value{

}
\description{
Title
}
