% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_net.R
\name{coexpression_modules_grnboost2}
\alias{coexpression_modules_grnboost2}
\title{Co-expression modules with GRNBoost2}
\usage{
coexpression_modules_grnboost2(GRANetObject, threads)
}
\arguments{
\item{GRANetObject}{Object created using the function CreateGRAnetObject}

\item{threads}{Number of threads to use.}
}
\value{
GRANetObject with computed co-expression modules.
}
\description{
Computes co-expression modules using GRNBoost2. This function uses the python
packages Arboreto and pySCENIC, please install them before (the vignette
01_create_condaenv can be follow to install all the python requirements).
The input is a GRANet object initialized from a Seurat object containing the
scRNA-seq data as counts, and a list of transcription factors as gene
symbols. The gene names in the original Seurat object should be symbols as
well.
}
\examples{
\dontrun{
granetobj <- coexpression_modules_grnboost2(GRANetObject = granetobj,
threads = 8)
}
}
