% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_net.R
\name{compute_coexpression_modules}
\alias{compute_coexpression_modules}
\title{Title}
\usage{
compute_coexpression_modules(GRANetObject, TFs, threads = 1)
}
\arguments{
\item{GRANetObject}{Object created using the function CreateGRAnetObject}

\item{TFs}{Character vector with the list of transcription factors included
in the Seurat object used to create the GRANetObject.}

\item{threads}{Number of threads to use.}
}
\value{

}
\description{
Title
}
