% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRANetMultiome_class.R
\name{CreateGRANetMultiomeObject}
\alias{CreateGRANetMultiomeObject}
\title{Creates a GRAnet object}
\usage{
CreateGRANetMultiomeObject(
  SeuratObject,
  RNA_assay = "RNA",
  ATAC_assay = "ATAC",
  motifSet,
  genome
)
}
\arguments{
\item{SeuratObject}{Seurat object with scRNA-seq counts and a categorical
cell identity annotation.}

\item{RNA_assay}{Name of the Assay in the input SeuratObject containing the
gene expression data.}

\item{ATAC_assay}{Name of the Assay in the input SeuratObject containing the
chromatin accessibility data.}

\item{motifSet}{Name of the motif set database to use. Supported motif sets
are: "jaspar2020", "jaspar2018", "jaspar2016", "cisbp", "encode", "homer".}

\item{genome}{Genome of the organism under study, supported genomes are:
"hg19", "hg38", "mm9", and "mm10".}
}
\value{

}
\description{
A GRANetMultiome object is initialized from a Seurat object containing scRNA-seq
data as counts. The gene names in the original count matrix should be gene
symbols as they are required to match to the motifs associated to a list of
transcription factors. Additionally, one of the columns of the cell metadata
from the Seurat object should contain a categorical annotation of the cells
(e.g., cluster, cell type, cell state).
}
\examples{
\dontrun{
granetobj <- CreateGRAnetObject(SeuratObject = seuratobj,
cssCluster = "tissue",
genome = "mm9")
}
}
