% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssRegulons_AUCell.R
\name{cssRegulons_activity}
\alias{cssRegulons_activity}
\title{Title}
\usage{
cssRegulons_activity(GRANetObject, aucMaxRank = "1\%", threads = 1)
}
\arguments{
\item{aucMaxRank}{possible values: min, 1%, 5%, 10%, 50%, 100%}

\item{threads}{}
}
\value{

}
\description{
Title
}
