% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRANet_class.R
\docType{class}
\name{GRANet-class}
\alias{GRANet-class}
\alias{GRANet}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{SeuratObject}}{Seurat.}

\item{\code{cssCluster}}{character.}

\item{\code{ProjectMetadata}}{list.}

\item{\code{TFmotif_location}}{list.}

\item{\code{Coexprs_modules}}{data.frame.}

\item{\code{cssRegulons}}{list.}

\item{\code{cssRegulonsAUCell}}{matrix.}

\item{\code{.}}{}
}}

