% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRANetMultiome_class.R
\docType{class}
\name{GRANetMultiome-class}
\alias{GRANetMultiome-class}
\alias{GRANetMultiome}
\title{Creates a new GRANet object}
\value{

}
\description{
Creates a new GRANet object
}
\section{Slots}{

\describe{
\item{\code{SeuratObject}}{Seurat.}

\item{\code{ProjectMetadata}}{list.}

\item{\code{TFmotif_location}}{list.}

\item{\code{Coexprs_modules}}{data.frame.}

\item{\code{cssRegulons}}{list.}

\item{\code{cssRegulonsAUCell}}{matrix.}
}}

