% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_hgnc_curated_tfs}
\alias{hs_hgnc_curated_tfs}
\title{Curated human transcription factors}
\format{
Curated list of symbols of human transcription factors
\describe{
  \item{hs_hgnc_curated_tfs}{1390 transcription factors symbols}
  ...
}
}
\source{
\url{https://github.com/aertslab/pySCENIC/tree/master/resources}
}
\usage{
data(hs_hgnc_curated_tfs)
}
\description{
List curated by Van de Sande, B., Flerin, C., Davie, K. et al.
A scalable SCENIC workflow for single-cell gene regulatory network analysis.
Nat Protoc 15, 2247–2276 (2020). https://doi.org/10.1038/s41596-020-0336-2
}
\keyword{datasets}
