% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_scCATseq_embryo}
\alias{hs_scCATseq_embryo}
\title{Human embryo scCAT-seq data}
\format{
List with gene and peak counts, annotation and Seurat object with gene counts as symbols.
\describe{
  \item{hs_scCATseq_embryo}{72 human cells from the blastocyst and morula stage}
  ...
}
}
\source{
\url{https://doi.org/10.1038/s41467-018-08205-7}
}
\usage{
data(hs_scCATseq_embryo)
}
\description{
Gene expression and chromatin accessibility data for 72 human cells from the
blastocyst and morula stage.
Liu, L., Liu, C., Quintero, A. et al.
Deconvolution of single-cell multi-omics layers reveals regulatory heterogeneity.
Nat Commun 10, 470 (2019). https://doi.org/10.1038/s41467-018-08205-7
}
\keyword{datasets}
