% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRANetMultiome_reg_interactions.R
\name{infer_reg_interactions}
\alias{infer_reg_interactions}
\title{Quantification of regulon activity and binarization}
\usage{
infer_reg_interactions(
  GRANetObject,
  min_cell_per_interaction = 0.01,
  test_p_val_threshold = 0.05,
  test_by_bins = TRUE,
  binsize = 1000
)
}
\arguments{
\item{GRANetObject}{GRANet object with computed regulons.}

\item{min_cell_per_interaction}{Minimum percentage of cells that are required
to have one regulatory interaction. Regulatory interactions found in fewer
cells than min_cell_per_interaction are filtered out.}

\item{test_p_val_threshold}{p-value threshold in order ot keep only regulatory
interactions that are related with a change in the target gene expression.}
}
\value{

}
\description{
Quantification of regulon activity and binarization
}
\examples{
\dontrun{
granetobj <- infer_reg_interactions(GRANetObject=granetobj,
min_cell_per_interaction = 0.01,
test_p_val_threshold     = 0.05)
}
}
