% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lambert2018}
\alias{lambert2018}
\title{Human transcription factors from Lambert 2018}
\format{
List of symbols of human transcription factors from Lambert 2018
\describe{
  \item{lambert2018}{1639 transcription factors symbols}
  ...
}
}
\source{
\url{https://github.com/aertslab/pySCENIC/tree/master/resources}
}
\usage{
data(lambert2018)
}
\description{
List curated by Van de Sande, B., Flerin, C., Davie, K. et al.
A scalable SCENIC workflow for single-cell gene regulatory network analysis.
Nat Protoc 15, 2247–2276 (2020). https://doi.org/10.1038/s41596-020-0336-2
}
\details{
Originally published by Lambert SA, Jolma A, Campitelli LF, Das PK, Yin Y,
Albu M, Chen X, Taipale J, Hughes TR, Weirauch MT. The Human Transcription
Factors. Cell. 2018 Feb 8;172(4):650-665. doi: 10.1016/j.cell.2018.01.029.
Erratum in: Cell. 2018 Oct 4;175(2):598-599. PMID: 29425488.
}
\keyword{datasets}
