% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssRegulons.R
\name{make_cssRegulons}
\alias{make_cssRegulons}
\title{Construction of cell state-specific Regulons (cssRegulons)}
\usage{
make_cssRegulons(
  GRANetObject,
  promoter_size = 5000,
  min_regulon_size = 20,
  threads = 1
)
}
\arguments{
\item{GRANetObject}{GRANet object with computed co-expression modules and
extracted locations of motifs linked to open chromatin regions.}

\item{promoter_size}{Window size to search for transcription factor motifs
located around the TSS of all target genes from the co-expression modules.}

\item{min_regulon_size}{Minimum number of target genes required to build a
regulon, composed of a transcription factor and a list of cis-regulated
target genes.}

\item{threads}{Number of threads to use.}
}
\value{

}
\description{
Construction of cell state-specific Regulons (cssRegulons)
}
\examples{
\dontrun{
granetobj <- make_cssRegulons(GRANetObject = granetobj,
promoter_size = 50000,
min_regulon_size=20)
}
}
