% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRANetMultiome_match_and_trim.R
\name{match_motifs_and_trim}
\alias{match_motifs_and_trim}
\title{Find TFs motif positions linked to open chromatin regions and make regulons}
\usage{
match_motifs_and_trim(
  GRANetObject,
  promoter_size,
  min_regulon_size = 20,
  cutOff = 5e-05,
  width = 7
)
}
\arguments{
\item{GRANetObject}{GRANet object with computed co-expression modules and
Spearman correlation between transcription factors and target genes.}

\item{promoter_size}{Window size to search for transcription factor motifs
located around the TSS of all target genes from the co-expression modules.}

\item{min_regulon_size}{Minimum number of target genes required to build a
regulon, composed of a transcription factor and a list of cis-regulated
target genes.}

\item{cutOff}{motifmatchr.: p-value cutoff for returning motifs}

\item{width}{motifmatchr: p-value cutoff for returning motifs}
}
\value{

}
\description{
This function takes as input a GRANetMultiome object,
and it extracts the position of motifs associated to transcription factors
linked to open chromatin regions.
All the positions of transcription factors motifs included in the co-expression
modules stored in the GRANetMultiome object are searched using motifmatchr.
Then the cis-regulatory interactions inside the selected promoter size, are
kept in order to create the regulons.
}
\examples{
\dontrun{
granetobj <- match_motifs_and_trim(GRANetObject=granetobj,
promoter_size=10000)
}
}
