% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm_mgi_tfs}
\alias{mm_mgi_tfs}
\title{Mouse transcription factors}
\format{
List of symbols of mouse transcription factors
\describe{
  \item{mm_mgi_tfs}{1721 transcription factors symbols}
  ...
}
}
\source{
\url{https://github.com/aertslab/pySCENIC/tree/master/resources}
}
\usage{
data(mm_mgi_tfs)
}
\description{
List curated by Van de Sande, B., Flerin, C., Davie, K. et al.
A scalable SCENIC workflow for single-cell gene regulatory network analysis.
Nat Protoc 15, 2247–2276 (2020). https://doi.org/10.1038/s41596-020-0336-2
}
\keyword{datasets}
