% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRANetMultiome_regulon_AUCell.R
\name{regulons_activity}
\alias{regulons_activity}
\title{Quantification of regulon activity and binarization}
\usage{
regulons_activity(GRANetObject, aucMaxRank = "1\%", threads = 1)
}
\arguments{
\item{GRANetObject}{GRANet object with computed regulons.}

\item{aucMaxRank}{possible values: "min", "1\%", "5\%", "10\%", "50\%", "100\%".}

\item{threads}{Number of threads to use.}
}
\value{

}
\description{
Quantification of regulon activity and binarization
}
\examples{
\dontrun{
granetobj <- regulons_activity(GRANetObject=granetobj,
aucMaxRank="50\%", threads=1)
}
}
