% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_net.R
\name{compute_coexpression_modules}
\alias{compute_coexpression_modules}
\title{Title}
\usage{
compute_coexpression_modules(GRANetObject, TFs, threads = 1)
}
\arguments{
\item{GRANetObject}{}
}
\value{

}
\description{
Title
}
