% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_net.R
\name{compute_coexpression_modules}
\alias{compute_coexpression_modules}
\title{Co-expression modules with GRNBoost2}
\usage{
compute_coexpression_modules(GRANetObject, TFs, threads = 1)
}
\arguments{
\item{GRANetObject}{Object created using the function CreateGRAnetObject}

\item{TFs}{Character vector with the list of transcription factors included
in the Seurat object used to create the GRANetObject.}

\item{threads}{Number of threads to use.}
}
\value{
GRANetObject with computed co-expression modules.
}
\description{
Computes co-expression modules using GRNBoost2. This function uses the python
packages Arboreto and pySCENIC, please install them before (the vignette
01_create_condaenv can be follow to install all the python requirements).
The input is a GRANet object initialized from a Seurat object containing the
scRNA-seq data as counts, and a list of transcription factors as gene
symbols. The gene names in the original Seurat object should be symbols as
well.
}
\examples{
\dontrun{
granetobj <- compute_coexpression_modules(GRANetObject = granetobj,
TFs = readLines("data/mm_mgi_tfs.txt"),
threads = 8)
}
}
