% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssRegulons_AUCell.R
\name{cssRegulons_activity}
\alias{cssRegulons_activity}
\title{Quantification of cssRegulon activity}
\usage{
cssRegulons_activity(GRANetObject, aucMaxRank = "1\%", threads = 1)
}
\arguments{
\item{GRANetObject}{GRANet object with computed cssRegulons.}

\item{aucMaxRank}{possible values: "min", "1\%", "5\%", "10\%", "50\%", "100\%".}

\item{threads}{Number of threads to use.}
}
\value{

}
\description{
Quantification of cssRegulon activity
}
\examples{
\dontrun{
granetobj <- cssRegulons_activity(GRANetObject=granetobj,
aucMaxRank="50\%", threads=1)
}
}
