% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssRegulons.R
\name{make_cssRegulons}
\alias{make_cssRegulons}
\title{Title}
\usage{
make_cssRegulons(
  GRANetObject,
  promoter_size = 5000,
  min_regulon_size = 20,
  threads = 1
)
}
\arguments{
\item{threads}{}
}
\value{

}
\description{
Title
}
